Minimizes gene looping
