+++ scout/server/blueprints/variant/utils.py
@@ -5,28 +5,278 @@ from scout.server.links import (add_gene_links, ensembl, add_tx_links)
 
 LOG = logging.getLogger(__name__)
 
-def transcript_str(transcript_obj, gene_name=None):
-    """Generate amino acid change as a string."""
-    if transcript_obj.get('exon'):
-        gene_part, part_count_raw = 'exon', transcript_obj['exon']
-    elif transcript_obj.get('intron'):
-        gene_part, part_count_raw = 'intron', transcript_obj['intron']
-    else:
-        # variant between genes
-        gene_part, part_count_raw = 'intergenic', '0'
+def add_panel_specific_gene_info(panel_info, gene_obj):
+    """Adds manualy curated information from a gene panel to a gene
+    
+    Args:
+        panel_info(list(dict)): List of panel information about a gene
+        gene_obj(dict): The gene information from the variant
+    
+    Returns:
+        panel_specific(dict): A dictionary with a summary of the information from gene panels
+    """
+    panel_specific = {}
+    # Manually annotated disease associated transcripts
+    disease_associated = set()
+    # We need to strip the version to compare against others
+    disease_associated_no_version = set()
+    manual_penetrance = False
+    mosaicism = False
+    manual_inheritance = set()
 
-    part_count = part_count_raw.rpartition('/')[0]
-    change_str = "{}:{}{}:{}:{}".format(
-        transcript_obj.get('refseq_id', ''),
-        gene_part,
-        part_count,
-        transcript_obj.get('coding_sequence_name', 'NA'),
-        transcript_obj.get('protein_sequence_name', 'NA'),
-    )
-    if gene_name:
-        change_str = "{}:".format(gene_name) + change_str
-    return change_str
+    # We need to loop since there can be information from multiple panels
+    for gene_info in panel_info:
+        # Check if there are manually annotated disease transcripts
+        for tx in gene_info.get('disease_associated_transcripts', []):
+            # We remove the version of transcript at this stage
+            stripped = re.sub(r'\.[0-9]', '', tx)
+            disease_associated_no_version.add(stripped)
+            disease_associated.add(tx)
 
+        if gene_info.get('reduced_penetrance'):
+            manual_penetrance = True
+
+        if gene_info.get('mosaicism'):
+            mosaicism = True
+
+        manual_inheritance.update(gene_info.get('inheritance_models', []))
+
+    panel_specific['disease_associated_transcripts'] = list(disease_associated)
+    panel_specific['disease_associated_no_version'] = disease_associated_no_version
+    panel_specific['manual_penetrance'] = manual_penetrance
+    panel_specific['mosaicism'] = mosaicism
+    panel_specific['manual_inheritance'] = list(manual_inheritance)
+    
+    return panel_specific
+
+def update_transcripts_information(variant_gene, hgnc_gene, variant_obj, genome_build=None):
+    """Collect tx info from the hgnc gene and panels and update variant transcripts
+    
+    Since the hgnc information are continuously being updated we need to run this each time a 
+    variant is fetched.
+    
+    This function will:
+        - Add a dictionary with tx_id -> tx_info to the hgnc variant
+        - Add information from the panel
+    
+    Args:
+        variant_gene(dict): the gene information from the variant
+        hgnc_gene(dict): the hgnc gene information
+        varaiant_obj(scout.models.Variant)
+    
+    """
+    genome_build = genome_build or '37'
+    disease_associated_no_version = variant_gene.get('disease_associated_no_version', set())
+    # Create a dictionary with transcripts information
+    # Use ensembl transcript id as keys
+    transcripts_dict = {}
+    # Add transcript information from the hgnc gene
+    for transcript in hgnc_gene.get('transcripts', []):
+        tx_id = transcript['ensembl_transcript_id']
+        transcripts_dict[tx_id] = transcript
+
+    # Add the transcripts to the gene object
+    hgnc_gene['transcripts_dict'] = transcripts_dict
+    hgnc_symbol = hgnc_gene['hgnc_symbol']
+    refseq_transcripts = []
+
+    # First loop over the variants transcripts
+    for transcript in variant_gene.get('transcripts', []):
+        tx_id = transcript['transcript_id']
+        hgnc_transcript = transcripts_dict.get(tx_id)
+        # If the tx does not exist in ensembl anymore we skip it
+        if not hgnc_transcript:
+            continue
+
+        # Check in the common information if it is a primary transcript
+        if hgnc_transcript.get('is_primary'):
+            transcript['is_primary'] = True
+        
+        # Add the transcript links
+        add_tx_links(transcript, genome_build)
+        # If the transcript has a ref seq identifier we add that
+        # to the variants transcript
+        refseq_id = hgnc_transcript.get('refseq_id')
+        if not refseq_id:
+            continue
+        transcript['refseq_id'] = refseq_id
+        variant_obj['has_refseq'] = True
+        
+        refseq_transcripts.append(transcript)
+        # Check if the refseq id are disease associated
+        if refseq_id in disease_associated_no_version:
+            transcript['is_disease_associated'] = True
+
+        # Since a ensemble transcript can have multiple refseq identifiers we add all of
+        # those
+        transcript['refseq_identifiers'] = hgnc_transcript.get('refseq_identifiers',[])
+
+        transcript['change_str'] = transcript_str(transcript, hgnc_symbol)
+    
+    variant_gene['primary_transcripts'] = refseq_transcripts
+        
+
+def add_gene_info(store, variant_obj, gene_panels=None, genome_build=None):
+    """Add extra information about genes from gene panels.
+    
+    There are are sometimes additional information that are manually curated in the gene panels.
+    This information needs to be added to the variant before sending it to the template.
+
+    This function will loop over all genes and add that extra information.
+
+    Args:
+        store(scout.adapter.MongoAdapter)
+        variant_obj(dict): A variant from the database
+        gene_panels(list(dict)): List of panels from database
+        genome_build(str)
+    
+    Returns:
+        variant_obj
+    """
+    gene_panels = gene_panels or []
+    genome_build = genome_build or '37'
+
+    # Add a variable that checks if there are any refseq transcripts
+
+    # extra_info will hold information from gene panels
+    extra_info = {}
+    for panel_obj in gene_panels:
+        for gene_info in panel_obj['genes']:
+            hgnc_id = gene_info['hgnc_id']
+            if hgnc_id not in extra_info:
+                extra_info[hgnc_id] = []
+
+            extra_info[hgnc_id].append(gene_info)
+
+    # Loop over the genes in the variant object to add information
+    # from hgnc_genes and panel genes to the variant object
+    variant_obj['has_refseq'] = False
+    variant_obj['disease_associated_transcripts'] = []
+    all_models = set()
+    for variant_gene in variant_obj.get('genes', []):
+        hgnc_id = variant_gene['hgnc_id']
+        # Get the hgnc_gene
+        hgnc_gene = store.hgnc_gene(hgnc_id, build=genome_build)
+
+        if not hgnc_gene:
+            continue
+        hgnc_symbol = hgnc_gene['hgnc_symbol']
+        # Add omim information if gene is annotated to have incomplete penetrance
+        if hgnc_gene.get('incomplete_penetrance'):
+            variant_gene['omim_penetrance'] = True
+
+        ############# PANEL SPECIFIC INFORMATION #############
+        # Panels can have extra information about genes and transcripts
+        panel_info = add_panel_specific_gene_info(extra_info.get(hgnc_id, []), variant_gene)
+        variant_gene.update(panel_info)
+        
+        update_transcripts_information(variant_gene, hgnc_gene, variant_obj)
+        
+        variant_gene['common'] = hgnc_gene
+        
+        add_gene_links(variant_gene, genome_build)
+        
+        # Add disease associated transcripts to variant
+        for refseq_id in panel_info.get('disease_associated_transcripts', []):
+            transcript_str = "{}:{}".format(hgnc_symbol, refseq_id)
+            variant_obj['disease_associated_transcripts'].append(transcript_str)
+        
+        # Add the associated disease terms
+        disease_terms = store.disease_terms(hgnc_id)
+        variant_gene['disease_terms'] = disease_terms
+
+        all_models = all_models.union(set(variant_gene['manual_inheritance']))
+        omim_models = set()
+        for disease_term in variant_gene.get('disease_terms', []):
+            omim_models.update(disease_term.get('inheritance', []))
+        variant_gene['omim_inheritance'] = list(omim_models)
+
+        all_models = all_models.union(omim_models)
+    
+    variant_obj['all_models'] = all_models
+
+    return variant_obj
+
+def predictions(genes):
+    """Adds information from variant specific genes to display.
+    
+    Args:
+        genes(list[dict])
+    
+    Returns:
+        data(dict)
+    """
+    data = {
+        'sift_predictions': [],
+        'polyphen_predictions': [],
+        'region_annotations': [],
+        'functional_annotations': []
+    }
+    for gene_obj in genes:
+        for pred_key in data:
+            gene_key = pred_key[:-1]
+            if len(genes) == 1:
+                value = gene_obj.get(gene_key, '-')
+            else:
+                gene_id = gene_obj.get('hgnc_symbol') or str(gene_obj['hgnc_id'])
+                value = ':'.join([gene_id, gene_obj.get(gene_key, '-')])
+            data[pred_key].append(value)
+
+    return data
+
+def sv_frequencies(variant_obj):
+    """Add frequencies in the correct way for the template
+    
+    Args:
+        variant_obj(scout.models.Variant)
+    
+    Returns:
+        frequencies(list(tuple)): A list of frequencies to display
+    """
+    
+    # Mandatory frequencies
+    
+    frequencies = [
+        ('GnomAD', variant_obj.get('gnomad_frequency')),
+    ]
+
+    if 'clingen_cgh_benign' in variant_obj:
+        frequencies.append(('ClinGen CGH (benign)', variant_obj['clingen_cgh_benign']))
+    if 'clingen_cgh_pathogenic' in variant_obj:
+        frequencies.append(('ClinGen CGH (pathogenic)', variant_obj['clingen_cgh_pathogenic']))
+    if 'clingen_ngi' in variant_obj:
+        frequencies.append(('ClinGen NGI', variant_obj['clingen_ngi']))
+    if 'clingen_mip' in variant_obj:
+        frequencies.append(('ClinGen MIP', variant_obj['clingen_mip']))
+    if 'swegen' in variant_obj:
+        frequencies.append(('SweGen', variant_obj['swegen']))
+    if 'decipher' in variant_obj:
+        frequencies.append(('Decipher', variant_obj['decipher']))
+    
+    return frequencies
+
+
+def is_affected(variant_obj, case_obj):
+    """Adds information to show in view if sample is affected
+    
+    The views sometimes expect strings so we need to convert the raw data. This is an typical 
+    example of that situation
+    
+    Loop over the samples in a variant and add information from the case is they are affected
+    
+    Args:
+        variant_obj(scout.models.Variant)
+    """
+    individuals = {individual['individual_id']: individual for individual in
+                   case_obj['individuals']}
+    for sample_obj in variant_obj['samples']:
+        individual = individuals[sample_obj.get('sample_id')]
+        if not individual:
+            continue
+        sample_obj['is_affected'] = False 
+        
+        if individual['phenotype'] == 2:
+            sample_obj['is_affected'] = True 
 
 def evaluation(store, evaluation_obj):
     """Fetch and fill-in evaluation object."""
@@ -38,33 +288,36 @@ def evaluation(store, evaluation_obj):
     evaluation_obj['classification'] = ACMG_COMPLETE_MAP.get(evaluation_obj['classification'])
     return evaluation_obj
 
-def parse_gene(gene_obj, build=None):
-    """Parse variant genes."""
-    build = build or 37
+def transcript_str(transcript_obj, gene_name=None):
+    """Generate amino acid change as a string.
+    
+    Args:
+        transcript_obj(dict)
+        gene_name(str)
+    
+    Returns:
+        change_str(str): A description of the transcript level change
+    """
+    # variant between genes
+    gene_part, part_count_raw = 'intergenic', '0'
 
-    if gene_obj.get('common'):
-        add_gene_links(gene_obj, build)
-        refseq_transcripts = []
-        for tx_obj in gene_obj['transcripts']:
-            parse_transcript(gene_obj, tx_obj, build)
+    if transcript_obj.get('exon'):
+        gene_part, part_count_raw = 'exon', transcript_obj['exon']
+    elif transcript_obj.get('intron'):
+        gene_part, part_count_raw = 'intron', transcript_obj['intron']
 
-            # select refseq transcripts as "primary"
-            if not tx_obj.get('refseq_id'):
-                continue
+    part_count = part_count_raw.rpartition('/')[0]
+    change_str = "{}:{}{}:{}:{}".format(
+        transcript_obj.get('refseq_id', ''),
+        gene_part,
+        part_count,
+        transcript_obj.get('coding_sequence_name', 'NA'),
+        transcript_obj.get('protein_sequence_name', 'NA'),
+    )
+    if gene_name:
+        change_str = "{}:".format(gene_name) + change_str
 
-            refseq_transcripts.append(tx_obj)
-
-        gene_obj['primary_transcripts'] = (refseq_transcripts if refseq_transcripts else [])
-
-def parse_transcript(gene_obj, tx_obj, build=None):
-    """Parse variant gene transcript (VEP)."""
-    build = build or 37
-    add_tx_links(tx_obj, build)
-
-    if tx_obj.get('refseq_id'):
-        gene_name = (gene_obj['common']['hgnc_symbol'] if gene_obj['common'] else
-                     gene_obj['hgnc_id'])
-        tx_obj['change_str'] = transcript_str(tx_obj, gene_name)
+    return change_str
 
 
 def frequency(variant_obj):
